Add files via upload
Latest search function, utilizing json-server to local host the json. Event listeners have effectively been added to track the activity of the checkboxes on the site, however Ill need to review the effectiveness of the search blocks of code triggered by the checkboxed. 
The uploaded files, searchsk1(the html and js file) are meant to be the skeletons of the ultimate search function however features that will be included later include:
-only being able to select one checkbox at a time
-including keywords attribute in the search
-adding job family search function which still needs to be defined
-currently debugging the current filter checkboxes
//if any errors are in the writing, I'm willing to clarify in person.